+++ src/airport/model/Airport.java
@@ -1,4 +1,4 @@
-package airport;
+package airport.model;
 
 import java.util.ArrayList;
 import java.util.List;
